# Euclidean Specialization for ArgKminClassMode