@jjerphan 2024-01-22
Thank you, @OmarManzoor.

I do not have time to have a look into it at the moment. IIRC, there are two pathways for this case (extend `EuclideanArgKmin` or extend `ArgKminClassMode`).

Can you report benchmarks for various configurations against `main`, please?