@OmarManzoor 2024-01-23
HI @jjerphan 

I did run some benchmarks. It seems that when we have test samples of about a 100 the performance has decreased. Otherwise for large test sizes the performance seems to have increased.

```console
· Running 2 total benchmarks (2 commits * 1 environments * 1 benchmarks)
[ 0.00%] · For scikit-learn commit 5d0219cf <argkmin_classmode_euclidean> (round 1/1):
[ 0.00%] ·· Benchmarking conda-py3.10-cython3.0.3-joblib1.3.2-numpy1.25.2-pandas2.1.0-scipy1.11.2-threadpoolctl3.2.0
[50.00%] ··· argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba                                                                                                                                                ok
[50.00%] ··· ========= ======== ============= ============ ============ ============= ============= ============ ============ ============ ============
             --                                                                 n_features / n_neighbors                                               
             ------------------ -----------------------------------------------------------------------------------------------------------------------
              n_train   n_test     50 / 100     50 / 500    50 / 1000     100 / 100     100 / 500    100 / 1000   500 / 100    500 / 500    500 / 1000 
             ========= ======== ============= ============ ============ ============= ============= ============ ============ ============ ============
                1000     100      2.24±0.2ms   4.12±0.3ms   3.42±0.3ms    2.27±0.2ms    4.38±0.3ms   3.76±0.4ms   4.02±0.5ms    6.25±1ms    5.89±0.7ms 
                1000     1000    4.13±0.02ms    11.7±1ms    9.06±0.1ms   4.72±0.01ms   13.1±0.09ms   12.8±0.1ms   13.4±0.2ms   19.9±0.2ms   19.8±0.2ms 
                1000    10000      32.9±4ms     65.0±4ms     69.2±5ms      33.6±3ms      70.5±3ms     70.3±3ms     73.7±4ms     105±3ms      118±6ms   
               10000     100       15.4±4ms     20.1±2ms     28.7±2ms      14.6±1ms      23.1±2ms     32.2±3ms     39.1±4ms     45.6±4ms     58.7±6ms  
               10000     1000    15.5±0.03ms   33.2±0.4ms   50.2±0.1ms    21.5±0.1ms    39.4±0.1ms   55.9±0.1ms   68.3±0.5ms   87.3±0.4ms   106±0.9ms  
               10000    10000      119±5ms      243±4ms      373±7ms       175±4ms       301±6ms      446±10ms     639±20ms     773±20ms     937±30ms  
             ========= ======== ============= ============ ============ ============= ============= ============ ============ ============ ============

[50.00%] · For scikit-learn commit 8734d2f0 (round 1/1):
[50.00%] ·· Building for conda-py3.10-cython3.0.3-joblib1.3.2-numpy1.25.2-pandas2.1.0-scipy1.11.2-threadpoolctl3.2.0..
[50.00%] ·· Benchmarking conda-py3.10-cython3.0.3-joblib1.3.2-numpy1.25.2-pandas2.1.0-scipy1.11.2-threadpoolctl3.2.0
[100.00%] ··· argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba                                                                                                                                                ok
[100.00%] ··· ========= ======== ============= ============= ============= ============= ============= ============= ============= ============= ============
              --                                                                   n_features / n_neighbors                                                  
              ------------------ ----------------------------------------------------------------------------------------------------------------------------
               n_train   n_test     50 / 100      50 / 500     50 / 1000     100 / 100     100 / 500     100 / 1000    500 / 100     500 / 500    500 / 1000 
              ========= ======== ============= ============= ============= ============= ============= ============= ============= ============= ============
                 1000     100     1.60±0.01ms   3.62±0.06ms   5.61±0.09ms   1.66±0.01ms   3.70±0.05ms   5.70±0.09ms   2.22±0.02ms   4.33±0.07ms   6.39±0.1ms 
                 1000     1000     10.8±0.2ms     27.6±9ms     39.9±0.5ms    11.0±0.1ms    26.9±0.6ms    40.4±0.4ms    16.1±0.2ms    32.1±0.5ms   45.8±0.6ms 
                 1000    10000      42.8±1ms      152±4ms       290±7ms       48.5±2ms      156±3ms       297±8ms       88.7±2ms      197±6ms      324±5ms   
                10000     100      3.94±0.1ms    10.0±0.5ms    14.3±0.6ms    4.93±0.3ms    10.8±0.4ms    14.8±0.7ms    10.1±0.3ms    16.3±0.6ms    20.4±1ms  
                10000     1000      30.6±1ms      86.7±3ms      117±4ms       36.6±1ms      92.4±3ms      123±4ms       83.4±2ms      141±5ms      173±7ms   
                10000    10000      136±3ms       333±6ms       591±10ms      192±7ms       395±9ms       649±20ms      639±10ms      878±20ms    1.14±0.01s 
              ========= ======== ============= ============= ============= ============= ============= ============= ============= ============= ============

| Change   | Before [8734d2f0]    | After [5d0219cf] <argkmin_classmode_euclidean>   |   Ratio | Benchmark (Parameter)                                                                        |
|----------|----------------------|--------------------------------------------------|---------|----------------------------------------------------------------------------------------------|
| +        | 3.94±0.1ms           | 15.4±4ms                                         |    3.89 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 100, 50, 100)     |
| +        | 10.1±0.3ms           | 39.1±4ms                                         |    3.87 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 100, 500, 100)    |
| +        | 4.93±0.3ms           | 14.6±1ms                                         |    2.95 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 100, 100, 100)    |
| +        | 20.4±1ms             | 58.7±6ms                                         |    2.88 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 100, 500, 1000)   |
| +        | 16.3±0.6ms           | 45.6±4ms                                         |    2.8  | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 100, 500, 500)    |
| +        | 14.8±0.7ms           | 32.2±3ms                                         |    2.18 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 100, 100, 1000)   |
| +        | 10.8±0.4ms           | 23.1±2ms                                         |    2.13 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 100, 100, 500)    |
| +        | 14.3±0.6ms           | 28.7±2ms                                         |    2.01 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 100, 50, 1000)    |
| +        | 10.0±0.5ms           | 20.1±2ms                                         |    2.01 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 100, 50, 500)     |
| +        | 2.22±0.02ms          | 4.02±0.5ms                                       |    1.81 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 100, 500, 100)     |
| +        | 4.33±0.07ms          | 6.25±1ms                                         |    1.44 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 100, 500, 500)     |
| +        | 1.60±0.01ms          | 2.24±0.2ms                                       |    1.4  | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 100, 50, 100)      |
| +        | 1.66±0.01ms          | 2.27±0.2ms                                       |    1.37 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 100, 100, 100)     |
| +        | 3.70±0.05ms          | 4.38±0.3ms                                       |    1.18 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 100, 100, 500)     |
| +        | 3.62±0.06ms          | 4.12±0.3ms                                       |    1.14 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 100, 50, 500)      |
| -        | 136±3ms              | 119±5ms                                          |    0.88 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 10000, 50, 100)   |
| -        | 878±20ms             | 773±20ms                                         |    0.88 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 10000, 500, 500)  |
| -        | 16.1±0.2ms           | 13.4±0.2ms                                       |    0.84 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 1000, 500, 100)    |
| -        | 88.7±2ms             | 73.7±4ms                                         |    0.83 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 10000, 500, 100)   |
| -        | 83.4±2ms             | 68.3±0.5ms                                       |    0.82 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 1000, 500, 100)   |
| -        | 1.14±0.01s           | 937±30ms                                         |    0.82 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 10000, 500, 1000) |
| -        | 42.8±1ms             | 32.9±4ms                                         |    0.77 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 10000, 50, 100)    |
| -        | 395±9ms              | 301±6ms                                          |    0.76 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 10000, 100, 500)  |
| -        | 333±6ms              | 243±4ms                                          |    0.73 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 10000, 50, 500)   |
| -        | 48.5±2ms             | 33.6±3ms                                         |    0.69 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 10000, 100, 100)   |
| -        | 649±20ms             | 446±10ms                                         |    0.69 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 10000, 100, 1000) |
| -        | 5.70±0.09ms          | 3.76±0.4ms                                       |    0.66 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 100, 100, 1000)    |
| -        | 591±10ms             | 373±7ms                                          |    0.63 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 10000, 50, 1000)  |
| -        | 32.1±0.5ms           | 19.9±0.2ms                                       |    0.62 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 1000, 500, 500)    |
| -        | 141±5ms              | 87.3±0.4ms                                       |    0.62 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 1000, 500, 500)   |
| -        | 5.61±0.09ms          | 3.42±0.3ms                                       |    0.61 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 100, 50, 1000)     |
| -        | 173±7ms              | 106±0.9ms                                        |    0.61 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 1000, 500, 1000)  |
| -        | 36.6±1ms             | 21.5±0.1ms                                       |    0.59 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 1000, 100, 100)   |
| -        | 197±6ms              | 105±3ms                                          |    0.54 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 10000, 500, 500)   |
| -        | 30.6±1ms             | 15.5±0.03ms                                      |    0.51 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 1000, 50, 100)    |
| -        | 26.9±0.6ms           | 13.1±0.09ms                                      |    0.49 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 1000, 100, 500)    |
| -        | 156±3ms              | 70.5±3ms                                         |    0.45 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 10000, 100, 500)   |
| -        | 123±4ms              | 55.9±0.1ms                                       |    0.45 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 1000, 100, 1000)  |
| -        | 11.0±0.1ms           | 4.72±0.01ms                                      |    0.43 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 1000, 100, 100)    |
| -        | 45.8±0.6ms           | 19.8±0.2ms                                       |    0.43 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 1000, 500, 1000)   |
| -        | 152±4ms              | 65.0±4ms                                         |    0.43 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 10000, 50, 500)    |
| -        | 92.4±3ms             | 39.4±0.1ms                                       |    0.43 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 1000, 100, 500)   |
| -        | 117±4ms              | 50.2±0.1ms                                       |    0.43 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 1000, 50, 1000)   |
| -        | 27.6±9ms             | 11.7±1ms                                         |    0.42 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 1000, 50, 500)     |
| -        | 10.8±0.2ms           | 4.13±0.02ms                                      |    0.38 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 1000, 50, 100)     |
| -        | 86.7±3ms             | 33.2±0.4ms                                       |    0.38 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(10000, 1000, 50, 500)    |
| -        | 324±5ms              | 118±6ms                                          |    0.36 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 10000, 500, 1000)  |
| -        | 40.4±0.4ms           | 12.8±0.1ms                                       |    0.32 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 1000, 100, 1000)   |
| -        | 297±8ms              | 70.3±3ms                                         |    0.24 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 10000, 100, 1000)  |
| -        | 290±7ms              | 69.2±5ms                                         |    0.24 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 10000, 50, 1000)   |
| -        | 39.9±0.5ms           | 9.06±0.1ms                                       |    0.23 | argkmin_classmode.BruteForceKNeighborsClassifier.time_predict_proba(1000, 1000, 50, 1000)    |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.


```